--- a/src/services/api.ts
+++ b/src/services/api.ts
@@ -32,7 +32,7 @@ interface EstimateTradeAmountQuery {
   kind: OrderKind;
   amount: BigNumberish;
 }
-interface PlaceOrderQuery {
+export interface PlaceOrderQuery {
   order: Order;
   signature: Signature;
 }