--- a/src/tasks/decode.ts
+++ b/src/tasks/decode.ts
@@ -392,13 +392,20 @@ const setupDecodeTask: () => void = () => {
 
       displayTrades(trades, tokens, domainSeparator);
 
+      const tokenRegistry: Record<string, Token> = {};
+      tokens.forEach((token) => {
+        tokenRegistry[token.address] = token;
+      });
       const detailedInteractions = (await Promise.all(
         interactions.map(
           async (interactionGroup) =>
             await Promise.all(
               interactionGroup.map(async (i) => ({
                 ...i,
-                decoded: await decodeInteraction(i),
+                decoded: await decodeInteraction(i, hre, {
+                  tokenRegistry,
+                  settlementContractAddress: deployment?.address,
+                }),
               })),
             ),
         ),