--- a/src/tasks/withdraw.ts
+++ b/src/tasks/withdraw.ts
@@ -25,7 +25,7 @@ import {
   formatUsdValue,
   formatTokenValue,
   appraise,
-} from "./withdraw/value";
+} from "./ts/value";
 
 const DAI_DECIMALS = 18;