Migrates all new code from v0.0.1 to `main`.

The vast majority of the changes in this code are programmatically generated, based on commits `d03c006e6f7e69f0eec56fb8e9a71f2c69d8fdcb^..3e2ac9597c95e77127a6f0a2424f468785814bce` and `b129440bc078ae499c3db723ef4c0c2bc3a696e7^..v0.0.1`. Manual intervention was only consequence of merge conflicts or a final cleanup with minimal changes to ensure that the code works (see [here](https://github.com/gnosis/gp-v2-contracts/compare/merge-v0.0.1-changes-to-main...fix-merge-v0.0.1-changes-to-main?expand=1) for a readable version of the final changes).

I plan to merge this PR with fast forward, which means that commit history will be kept. The commit messages are the same as the original commits, with the exception that `[v0.0.1]` was removed and `(ported from v0.0.1)` was appended at the end.
  
<details><summary>Here is a list of all instructions run to obtain the commits, including a short comment on how merge conflicts were handled.</summary>

```
$ git cherry-pick d03c006e6f7e69f0eec56fb8e9a71f2c69d8fdcb^..3e2ac9597c95e77127a6f0a2424f468785814bce
Auto-merging src/tasks/withdraw.ts
[merge-v0.0.1-changes-to-main 41ad209] Fix too many events in event filter for withdraw script (#705)
 Date: Wed Jun 23 09:18:38 2021 +0000
 1 file changed, 70 insertions(+), 9 deletions(-)
Auto-merging src/tasks/withdraw.ts
[merge-v0.0.1-changes-to-main 4783830] Faster token info retrieval in withdraw script (#706)
 Date: Thu Jun 24 14:01:46 2021 +0000
 1 file changed, 50 insertions(+), 15 deletions(-)
Auto-merging src/tasks/withdraw.ts
CONFLICT (content): Merge conflict in src/tasks/withdraw.ts
error: could not apply 3e2ac95... [v0.0.1] Use prices from GPv2 services (#709)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# two new imports, both accepted
$ git cherry-pick b129440bc078ae499c3db723ef4c0c2bc3a696e7^..v0.0.1
Auto-merging src/tasks/decode.ts
[merge-v0.0.1-changes-to-main d3617bd] Add interaction decoding infrastructure (#715)
 Date: Mon Jul 5 08:29:07 2021 +0000
 3 files changed, 127 insertions(+), 8 deletions(-)
 create mode 100644 src/tasks/decode/interaction.ts
 create mode 100644 src/tasks/decode/interaction/template.ts
Auto-merging src/tasks/decode.ts
[merge-v0.0.1-changes-to-main d7610e9] Add Uniswap-like decoder (#718)
 Date: Mon Jul 5 14:11:57 2021 +0000
 3 files changed, 160 insertions(+), 5 deletions(-)
 create mode 100644 src/tasks/decode/interaction/uniswap_like.ts
[merge-v0.0.1-changes-to-main 889ca2b] Add ERC20 decoder (#716)
 Date: Mon Jul 5 14:22:06 2021 +0000
 2 files changed, 152 insertions(+), 1 deletion(-)
 create mode 100644 src/tasks/decode/interaction/erc20.ts
Auto-merging src/tasks/withdraw.ts
[merge-v0.0.1-changes-to-main 3684e58] [v0.0.1] Explicit network type in tasks (and related bugfix) (#727)
 Date: Wed Jul 7 08:47:23 2021 +0000
 3 files changed, 60 insertions(+), 24 deletions(-)
Auto-merging src/tasks/withdraw.ts
[merge-v0.0.1-changes-to-main 680437b] Isolate logic for rate limiting requests (#730)
 Date: Fri Jul 9 10:34:11 2021 +0000
 2 files changed, 135 insertions(+), 92 deletions(-)
 create mode 100644 src/tasks/ts/rate_limits.ts
Auto-merging src/ts/settlement.ts
[merge-v0.0.1-changes-to-main 56ed36e] [v0.0.1] services api (#735)
 Date: Thu Jul 15 18:38:19 2021 +0000
 3 files changed, 213 insertions(+), 45 deletions(-)
 create mode 100644 src/services/api.ts
[merge-v0.0.1-changes-to-main 97a1f16] [v0.0.1] Set endpoint in api (#747)
 Date: Fri Jul 23 15:01:43 2021 +0000
 2 files changed, 42 insertions(+), 9 deletions(-)
Auto-merging src/tasks/withdraw.ts
CONFLICT (content): Merge conflict in src/tasks/withdraw.ts
Removing src/tasks/ts/erc20.ts
Auto-merging src/tasks/decode.ts
error: could not apply b3c377c... [v0.0.1] Manage native token as erc20 token in tasks (#734)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# again two imports, both accepted
$ git cherry-pick --continue 
[merge-v0.0.1-changes-to-main 57cdeb8] [v0.0.1] Manage native token as erc20 token in tasks (#734)
 Date: Mon Jul 26 07:18:36 2021 +0000
 8 files changed, 280 insertions(+), 76 deletions(-)
 delete mode 100644 src/tasks/ts/erc20.ts
 create mode 100644 src/tasks/ts/tokens.ts
 create mode 100644 test/tasks/ts/tokens.test.ts
[merge-v0.0.1-changes-to-main c94231a] Add getFeeAndQuote to api (#752)
 Date: Tue Jul 27 10:48:16 2021 +0000
 2 files changed, 93 insertions(+), 15 deletions(-)
Auto-merging src/ts/order.ts
CONFLICT (content): Merge conflict in src/ts/order.ts
Auto-merging src/tasks/withdraw.ts
Auto-merging src/tasks/index.ts
error: could not apply 28be73a... [v0.0.1] Add script to dump tokens (#736)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# conflict in `normalizeOrder`, I rewrote the function so that the spirit of both changes was included
$ git cherry-pick --continue 
[merge-v0.0.1-changes-to-main 6a30729] [v0.0.1] Add script to dump tokens (#736)
 Date: Tue Jul 27 11:43:37 2021 +0000
 5 files changed, 607 insertions(+), 4 deletions(-)
 create mode 100644 src/tasks/dump.ts
 rename src/tasks/{withdraw => ts}/value.ts (100%)
Auto-merging yarn.lock
CONFLICT (content): Merge conflict in yarn.lock
Auto-merging package.json
CONFLICT (content): Merge conflict in package.json
error: could not apply 4ec6c1f... [v0.0.1] Add unit tests for dump script (#742)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# only added new packages to current package.json. Rebuilt yarn.lock
$ git cherry-pick --continue 
[merge-v0.0.1-changes-to-main 6d5ff17] [v0.0.1] Add unit tests for dump script (#742)
 Date: Tue Jul 27 13:37:41 2021 +0000
 5 files changed, 766 insertions(+), 6 deletions(-)
 create mode 100644 test/tasks/dump.test.ts
[merge-v0.0.1-changes-to-main 634efb4] Use getFeeAndQuoteSell in dump script (#757)
 Date: Fri Jul 30 13:46:28 2021 +0000
 3 files changed, 118 insertions(+), 147 deletions(-)
Auto-merging src/tasks/withdraw.ts
CONFLICT (content): Merge conflict in src/tasks/withdraw.ts
error: could not apply 45435e0... [v0.0.1] Isolate task functions (#759)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# accepted current changes while then changing the prompt function to the new one
$ git cherry-pick --continue 
[merge-v0.0.1-changes-to-main 2623988] [v0.0.1] Isolate task functions (#759)
 Date: Mon Aug 9 08:46:49 2021 +0000
 3 files changed, 380 insertions(+), 278 deletions(-)
 create mode 100644 src/tasks/withdraw/traded_tokens.ts
Auto-merging src/tasks/withdraw.ts
CONFLICT (content): Merge conflict in src/tasks/withdraw.ts
error: could not apply c77eedb... [v0.0.1] Explicit usd reference token in withdraw tasks (#760)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# accepted both imports
$ git cherry-pick --continue 
[merge-v0.0.1-changes-to-main cc84b61] [v0.0.1] Explicit usd reference token in withdraw tasks (#760)
 Date: Mon Aug 9 09:35:16 2021 +0000
 2 files changed, 63 insertions(+), 43 deletions(-)
CONFLICT (modify/delete): test/e2e/orderRefunds.test.ts deleted in HEAD and modified in 4be8bda ([v0.0.1] Add tests for withdraw script (#761)). Version 4be8bda ([v0.0.1] Add tests for withdraw script (#761)) of test/e2e/orderRefunds.test.ts left in tree.
Auto-merging test/e2e/burnFees.test.ts
Auto-merging hardhat.config.ts
CONFLICT (content): Merge conflict in hardhat.config.ts
error: could not apply 4be8bda... [v0.0.1] Add tests for withdraw script (#761)
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
# kept the variable `initialBaseFeePerGas` where it was and the order refund test deleted (note: I did it wrong and I had to rebase later to get rid of this file)
git cherry-pick --continue 
[merge-v0.0.1-changes-to-main 068d610] [v0.0.1] Add tests for withdraw script (#761)
 Date: Mon Aug 9 16:11:13 2021 +0000
 5 files changed, 436 insertions(+), 2 deletions(-)
 create mode 100644 src/contracts/test/TestERC20.sol
 create mode 100644 test/e2e/orderRefunds.test.ts
 create mode 100644 test/tasks/logging.ts
 create mode 100644 test/tasks/withdraw.test.ts
Auto-merging src/tasks/withdraw.ts
Auto-merging src/tasks/index.ts
[merge-v0.0.1-changes-to-main 2ba4d28] [v0.0.1] Add script for withdraw service (#762)
 Date: Wed Aug 11 10:32:48 2021 +0000
 7 files changed, 764 insertions(+), 5 deletions(-)
 create mode 100644 src/tasks/withdrawService.ts
 create mode 100644 test/tasks/withdrawService.test.ts
[merge-v0.0.1-changes-to-main 9cc5727] [v0.0.1] Withdraw service test pagination (#776)
 Date: Mon Aug 23 08:17:04 2021 +0000
 2 files changed, 200 insertions(+), 24 deletions(-)

$ git clone . ../gp-v2-contracts-surgery --no-local
$ cd ../gp-v2-contracts-surgery
$ git filter-repo --message-callback 'return re.sub(b"^(\[[vV]0.0.1\] )?([^\n]*)(\n.*)?$", b"\\2 (ported from v0.0.1)\\3", message, flags=re.DOTALL)' --refs origin/main..HEAD
Parsed 20 commits
New history written in 0.08 seconds...
HEAD is now at 8fb9054 Withdraw service test pagination (#776) (ported from v0.0.1)
Completely finished after 0.17 seconds.
$ cd ../gp-v2-contracts
$ git remote add local-surgery ../gp-v2-contracts-surgery
$ git fetch local-surgery
$ git reset --hard local-surgery/merge-v0.0.1-changes-to-main
```
</details>

The changes are rebased before merging to be able to fast forward on master.

### Test Plan

Except for CI, I ran the following manual checks.

1. Tried the withdraw script.
<details><summary>$ npx hardhat withdraw --network rinkeby --receiver 0x9008d19f58aabd9ed0d60971565aa8510560ab41 --min-value 10000 --leftover 500 --dry-run</summary>

```
Using account 0x00D374731083492b38A66ef5a1157cD378d020fc
Current account is not a solver. Only a solver can withdraw funds from the settlement contract.
Recovering list of traded tokens...
Ignored 14.497194445789504573 units of TUSD (0x0000000000085d4780B73119b644AE5ecd22b376) with value 70.83 USD
Ignored 44.206352312774136726 units of OWL (0xa7D1C04fAF998F9161fC9F800a99A809b84cfc9D) with value 2248.11 USD
Ignored 3.473963456574317576 units of PAX (0xBD6A9921504fae42EaD2024F43305A8ED3890F6f) with value 599.16 USD
Ignored 0.003196032974384109 units of UNI (0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984) with value 696.17 USD
Ignored 0.00000624 units of WBTC (0x577D296678535e4903D59A4C929B718e1D575e0A) with value 0.02 USD
Ignored 0.000241797368223218 units of BAT (0xbF7A7169562078c96f0eC1A8aFD6aE50f12e5A99) with value 7.85 USD
Ignored 0.000071500987578998 units of MKR (0xF9bA5210F91D0474bd1e1DcDAeC4C58E359AaD85) with value 259.19 USD
Amounts to withdraw:
 address                                    | balance (usd) | balance                        | withdrawn amount               | symbol 
--------------------------------------------+---------------+--------------------------------+--------------------------------+--------
 0x01BE23585060835E02B77ef475b0Cc51aA1e0709 |      11964.31 |           2.296388445927488155 |           2.200420241602840296 | LINK   
 0xddea378A6dDC8AfeC82C36E9b0078826bf9e68B6 |      18979.12 |           0.022091551110410854 |           0.021509555047803149 | ZRX    
 0xcFc99E8139Ae41FF6Db9E521f9e7cEaC8BDC9cbC |      41125.84 |           0.999911820374250438 |           0.987755087921692092 | TKIn   
 0xc7AD46e0b8a400Bb3C915120d284AafbA8fc4735 |      51142.65 | 4067972243939.2675003189347... | 4028201408915.4168251624184... | DAI    
 0xc778417E063141139Fce010982780140Aa0cD5Ab |     119353.34 |           1.984002693202415518 |           1.975691226525758606 | WETH   
 0xd0Dab4E640D95E9E8A47545598c33e31bDb53C7c |     166681.52 |         173.565375586339777772 |         173.044725890698985729 | GNO    
 0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa |    2234768.41 |     2234768.418125632329567535 |     2234268.418125632329567535 | DAI    
 0x4DBCdF9B62e891a7cec5A2568C3F4FAF9E8Abe2b |   53047863.09 |      307436.078017000000000000 |      307433.180293000000000000 | USDC   

The transaction will cost approximately 0.000171708002232204 ETH and will withdraw the balance of 8 tokens for an estimated total value of 55687878.32 USD. All withdrawn funds will be sent to 0x9008D19f58AAbD9eD0D60971565AA8510560ab41.
```
</details>

2. Tested the API with the (adapted) dedicated script from the related PR 
<details><summary>./buySomeTokens.ts</summary>

```
import { BigNumber, utils } from "ethers";
import hre from "hardhat";

import { Api, Environment } from "./src/services/api";
import { getDeployedContract } from "./src/tasks/ts/deployment";
import { domain, Order, OrderKind, SigningScheme, signOrder } from "./src/ts";

const sellToken = "0xc778417e063141139fce010982780140aa0cd5ab"; // weth

let tokens =
  "0x577D296678535e4903D59A4C929B718e1D575e0A 0xbF7A7169562078c96f0eC1A8aFD6aE50f12e5A99 0x6ED8111cA8779935d5e65A5dCB61872A386BE4A7 0xA6Cc591f2Fd8784DD789De34Ae7307d223Ca3dDc 0xD9BA894E0097f8cC2BBc9D24D308b98e36dc6D02 0x01BE23585060835E02B77ef475b0Cc51aA1e0709 0x4DBCdF9B62e891a7cec5A2568C3F4FAF9E8Abe2b 0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa 0xa7D1C04fAF998F9161fC9F800a99A809b84cfc9D 0xd0Dab4E640D95E9E8A47545598c33e31bDb53C7c 0xF9bA5210F91D0474bd1e1DcDAeC4C58E359AaD85 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984 0xc7AD46e0b8a400Bb3C915120d284AafbA8fc4735".split(
    " ",
  );
tokens = ["0xbadbadbadbadbadbadbadbadbadbadbadbadbadd", sellToken, ...tokens];

const kind = OrderKind.SELL;
const network = "rinkeby";
const chainId = 4;
if (hre.network.name !== network) {
  throw new Error("Bad network");
}

async function main() {
  const api = new Api(network, Environment.Dev);
  for (const buyToken of tokens) {
    console.log(`Token ${buyToken}`);
    const maxRand = 2000;
    const random = Math.floor(Math.random() * maxRand);
    const soldWeth = BigNumber.from(maxRand + random).mul(
      BigNumber.from(10).pow(18 - 5),
    );
    console.log(`Dumping ${utils.formatEther(soldWeth)} WETH`);
    let fee;
    try {
      fee = await api.getFee({
        sellToken,
        buyToken,
        kind,
        amount: soldWeth,
      });
    } catch (e) {
      console.log("Error fee:", (e as any).apiError);
      continue;
    }
    if (fee.gte(soldWeth)) {
      console.log("fee too large");
      continue;
    }
    const receivedAmount = await api.estimateTradeAmount({
      sellToken,
      buyToken,
      amount: soldWeth.sub(fee),
      kind,
    });
    const now = Math.floor(Date.now() / 1000);
    const order: Order = {
      sellToken,
      buyToken,
      sellAmount: soldWeth.sub(fee),
      buyAmount: receivedAmount,
      feeAmount: fee,
      kind: OrderKind.SELL,
      appData: "0x",
      validTo: now + 5 * 60,
      partiallyFillable: false,
    };

    const settlement = await getDeployedContract("GPv2Settlement", hre);
    const signer = (await hre.ethers.getSigners())[0];
    const domainSeparator = domain(chainId, settlement.address);
    const signature = await signOrder(
      domainSeparator,
      order,
      signer,
      SigningScheme.ETHSIGN,
    );

    console.log(`Creating order...`);
    try {
      console.log(
        "New order API returns:",
        await api.placeOrder({
          order,
          signature,
        }),
      );
    } catch (e) {
      console.log("Error create:", (e as any).apiError);
      continue;
    }
  }
}

main()
  .then(() => process.exit(0))
  .catch((error: any) => {
    console.error(error);
    process.exit(1);
  });
```
</details>

<details><summary>$ npx hardhat run ./buySomeTokens.ts --network rinkeby</summary>

```
Token 0xbadbadbadbadbadbadbadbadbadbadbadbadbadd
Dumping 0.02379 WETH
Error fee: { errorType: 'NoLiquidity', description: 'not enough liquidity' }
Token 0xc778417e063141139fce010982780140aa0cd5ab
Dumping 0.03376 WETH
fee too large
Token 0x577D296678535e4903D59A4C929B718e1D575e0A
Dumping 0.02401 WETH
Creating order...
New order API returns: 0xf054265eda5318fd7e405a1194fb1857026117388e1f0defc0559e58786ed09659552a04d9f0c184cfce5f951fe36a01a2b4add96137ccef
Token 0xbF7A7169562078c96f0eC1A8aFD6aE50f12e5A99
Dumping 0.0243 WETH
Creating order...
New order API returns: 0xacf3da9caf65cab2904e2aa836c20a212747ee93a8977d10050ae1f8f590160059552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf1
Token 0x6ED8111cA8779935d5e65A5dCB61872A386BE4A7
Dumping 0.03847 WETH
Creating order...
New order API returns: 0x1a55209664efb7a186198f265ceb10349002a8d739722a98aabbb85c0ca1d15b59552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf2
Token 0xA6Cc591f2Fd8784DD789De34Ae7307d223Ca3dDc
Dumping 0.02239 WETH
Creating order...
New order API returns: 0x07c1016c2fa2952febd0484e2ba24e29e99f1d6d77bc534a3b58e51732c1753a59552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf2
Token 0xD9BA894E0097f8cC2BBc9D24D308b98e36dc6D02
Dumping 0.03905 WETH
Creating order...
New order API returns: 0xa0a513c5de16fc3c5d4a33115f1b5d49510c1d87a3a6ccb7cede2a4efb757ee959552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf2
Token 0x01BE23585060835E02B77ef475b0Cc51aA1e0709
Dumping 0.02326 WETH
Creating order...
New order API returns: 0x1128260ca158632948e8b400769a2f3861dfaf184651c8c2d4185978157fee8259552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf2
Token 0x4DBCdF9B62e891a7cec5A2568C3F4FAF9E8Abe2b
Dumping 0.0397 WETH
Creating order...
New order API returns: 0x25c3fbfbef0e9e174b945b61e366f0c65f3489fc5d10aae8aa5ace3870fa1eb259552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf3
Token 0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa
Dumping 0.03742 WETH
Creating order...
New order API returns: 0x9745e61423b588fa31ad6165b3bc6ffc1856fafa62a962f0a29a73267ab89b6559552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf3
Token 0xa7D1C04fAF998F9161fC9F800a99A809b84cfc9D
Dumping 0.03836 WETH
Creating order...
New order API returns: 0x795d9522831bef3f766bcf24e9226b56fd70456b23ba25d5e98db4def6b63cbb59552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf3
Token 0xd0Dab4E640D95E9E8A47545598c33e31bDb53C7c
Dumping 0.02597 WETH
Creating order...
New order API returns: 0x5e4886da9cbedb703c580f507d50dc255d47bf8637e041bc0153bc22604b6b9d59552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf4
Token 0xF9bA5210F91D0474bd1e1DcDAeC4C58E359AaD85
Dumping 0.03588 WETH
Creating order...
New order API returns: 0x3ec3e7c0ab77eb609c675f8c45b2960e107db63c85c5b86f07627191a3c23c0959552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf4
Token 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984
Dumping 0.03767 WETH
Creating order...
New order API returns: 0x280d2204e936bdbaa1e4117692105020d40a713050ee77c6b292582670fdc99159552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf5
Token 0xc7AD46e0b8a400Bb3C915120d284AafbA8fc4735
Dumping 0.02324 WETH
Creating order...
New order API returns: 0xb21f15b227c799f01952b0ef9cc9ccd2040b0f865e85a476d3adbf540b65c51659552a04d9f0c184cfce5f951fe36a01a2b4add96137ccf5
```
</details>

3. Tested the dump script with the proceeds from the previous script.
…